testing/requirements.txt
docker-pycreds==0.4.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
numpy==1.26.3
pandas==2.2.1
psycopg2==2.9.10
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2


